requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
numpy==2.2.2
PyYAML==6.0.2
google-cloud-aiplatform==1.82.0
google-cloud-bigquery==3.33.0
google-cloud-storage==2.19.0
